+++ packages/api-client/src/main/broadcast/index.ts
@@ -17,11 +17,4 @@
  *
  */
 
-import {Invitation} from '../invitation';
-
-interface InvitationList {
-  has_more: boolean;
-  invitations: Invitation[];
-}
-
-export {InvitationList};
+export * from './BroadcastAPI';