+++ packages/core/src/main/conversation/ConversationService.ts
@@ -24,6 +24,7 @@ import {
   NewConversation,
   NewOTRMessage,
   OTRRecipients,
+  UserClients,
 } from '@wireapp/api-client/dist/commonjs/conversation/';
 import {CONVERSATION_TYPING, ConversationMemberLeaveEvent} from '@wireapp/api-client/dist/commonjs/event/';
 import {StatusCode} from '@wireapp/api-client/dist/commonjs/http/';
@@ -41,7 +42,7 @@ import {
   PayloadBundleState,
   PayloadBundleType,
   ReactionType,
-} from '../conversation/root';
+} from '../conversation/';
 
 import {
   Article,
@@ -58,6 +59,7 @@ import {
   MessageDelete,
   MessageEdit,
   MessageHide,
+  Quote,
   Reaction,
   Text,
   Tweet,
@@ -85,8 +87,8 @@ import {
 
 import {TextContentBuilder} from './TextContentBuilder';
 
+import {CryptographyService, EncryptedAsset} from '../cryptography/';
 import * as AssetCryptography from '../cryptography/AssetCryptography.node';
-import {CryptographyService, EncryptedAsset} from '../cryptography/root';
 
 import {APIClient} from '@wireapp/api-client';
 
@@ -218,7 +220,7 @@ class ConversationService {
     payloadBundle: PayloadBundleOutgoingUnsent,
     userIds?: string[]
   ): Promise<PayloadBundleOutgoing> {
-    const {originalMessageId, text, linkPreviews, mentions} = payloadBundle.content as EditedTextContent;
+    const {originalMessageId, text, linkPreviews, mentions, quote} = payloadBundle.content as EditedTextContent;
 
     const textMessage = Text.create({content: text});
 
@@ -228,6 +230,13 @@ class ConversationService {
 
     if (mentions && mentions.length) {
       textMessage.mentions = mentions.map(mention => Mention.create(mention));
+    }
+
+    if (quote) {
+      textMessage.quote = Quote.create({
+        quotedMessageId: quote.quotedMessageId,
+        quotedMessageSha256: quote.quotedMessageSha256,
+      });
     }
 
     const editedMessage = MessageEdit.create({
@@ -475,6 +484,7 @@ class ConversationService {
     };
   }
 
+  // TODO: Move this to a generic "message sending class".
   private async sendOTRMessage(
     sendingClientId: string,
     conversationId: string,
@@ -495,28 +505,39 @@ class ConversationService {
     }
   }
 
-  private async onClientMismatch(
+  // TODO: Move this to a generic "message sending class" and make it private.
+  public async onClientMismatch(
     error: AxiosError,
     message: NewOTRMessage,
     plainTextArray: Uint8Array
   ): Promise<NewOTRMessage> {
     if (error.response && error.response.status === StatusCode.PRECONDITION_FAILED) {
-      const {missing: missingClients, deleted: deletedClients} = error.response.data;
+      const {missing, deleted}: {missing: UserClients; deleted: UserClients} = error.response.data;
 
-      if (Object.keys(deletedClients).length) {
-        for (const recipientId in message.recipients) {
-          for (const deletedClientId of deletedClients[recipientId]) {
-            delete message.recipients[recipientId][deletedClientId];
+      const deletedUserIds = Object.keys(deleted);
+      const missingUserIds = Object.keys(missing);
+
+      if (deletedUserIds.length) {
+        for (const deletedUserId of deletedUserIds) {
+          for (const deletedClientId of deleted[deletedUserId]) {
+            const deletedUser = message.recipients[deletedUserId];
+            if (deletedUser) {
+              delete deletedUser[deletedClientId];
+            }
           }
         }
       }
 
-      if (Object.keys(missingClients).length) {
-        const missingPreKeyBundles = await this.apiClient.user.api.postMultiPreKeyBundles(missingClients);
+      if (missingUserIds.length) {
+        const missingPreKeyBundles = await this.apiClient.user.api.postMultiPreKeyBundles(missing);
         const reEncryptedPayloads = await this.cryptographyService.encrypt(plainTextArray, missingPreKeyBundles);
-        for (const recipientId in message.recipients) {
-          for (const clientId in reEncryptedPayloads[recipientId]) {
-            message.recipients[recipientId][clientId] = reEncryptedPayloads[recipientId][clientId];
+        for (const missingUserId of missingUserIds) {
+          for (const missingClientId in reEncryptedPayloads[missingUserId]) {
+            const missingUser = message.recipients[missingUserId];
+            if (!missingUser) {
+              message.recipients[missingUserId] = {};
+            }
+            message.recipients[missingUserId][missingClientId] = reEncryptedPayloads[missingUserId][missingClientId];
           }
         }
       }
@@ -671,7 +692,7 @@ class ConversationService {
       state: PayloadBundleState.OUTGOING_SENT,
     };
 
-    const {text, linkPreviews, mentions} = payloadBundle.content as TextContent;
+    const {text, linkPreviews, mentions, quote} = payloadBundle.content as TextContent;
 
     const textMessage = Text.create({
       content: text,
@@ -683,6 +704,13 @@ class ConversationService {
 
     if (mentions && mentions.length) {
       textMessage.mentions = mentions.map(mention => Mention.create(mention));
+    }
+
+    if (quote) {
+      textMessage.quote = Quote.create({
+        quotedMessageId: quote.quotedMessageId,
+        quotedMessageSha256: quote.quotedMessageSha256,
+      });
     }
 
     let genericMessage = GenericMessage.create({
@@ -1070,6 +1098,10 @@ class ConversationService {
     });
   }
 
+  public getClientID(): string {
+    return this.clientID;
+  }
+
   public async addUser(conversationId: string, userId: string): Promise<string>;
   public async addUser(conversationId: string, userIds: string[]): Promise<string[]>;
   public async addUser(conversationId: string, userIds: string | string[]): Promise<string | string[]> {
@@ -1130,7 +1162,7 @@ class ConversationService {
     return this.apiClient.conversation.api.postTyping(conversationId, {status: CONVERSATION_TYPING.STOPPED});
   }
 
-  public setClientID(clientID: string) {
+  public setClientID(clientID: string): void {
     this.clientID = clientID;
   }