+++ packages/core/src/main/team/index.ts
@@ -17,7 +17,4 @@
  *
  */
 
-export * from './Invitation';
-export * from './InvitationAPI';
-export * from './InvitationList';
-export * from './InvitationRequest';
+export * from './TeamService';