+++ packages/cbor/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.0.94](https://github.com/wireapp/wire-web-packages/tree/master/packages/cbor/compare/@wireapp/cbor@3.0.93...@wireapp/cbor@3.0.94) (2018-10-23)
+
+**Note:** Version bump only for package @wireapp/cbor
+
+
+
+
+
+## [3.0.93](https://github.com/wireapp/wire-web-packages/tree/master/packages/cbor/compare/@wireapp/cbor@3.0.92...@wireapp/cbor@3.0.93) (2018-10-19)
+
+**Note:** Version bump only for package @wireapp/cbor
+
+
+
+
+
+## [3.0.92](https://github.com/wireapp/wire-web-packages/tree/master/packages/cbor/compare/@wireapp/cbor@3.0.91...@wireapp/cbor@3.0.92) (2018-10-17)
+
+**Note:** Version bump only for package @wireapp/cbor
+
+
+
+
+
+## [3.0.91](https://github.com/wireapp/wire-web-packages/tree/master/packages/cbor/compare/@wireapp/cbor@3.0.90...@wireapp/cbor@3.0.91) (2018-10-11)
+
+**Note:** Version bump only for package @wireapp/cbor
+
+
+
+
+
 ## [3.0.90](https://github.com/wireapp/wire-web-packages/tree/master/packages/cbor/compare/@wireapp/cbor@3.0.89...@wireapp/cbor@3.0.90) (2018-10-09)
 
 **Note:** Version bump only for package @wireapp/cbor