+++ packages/lru-cache/CHANGELOG.md
@@ -3,6 +3,38 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.0.9](https://github.com/wireapp/wire-web-packages/tree/master/packages/lru-cache/compare/@wireapp/lru-cache@3.0.8...@wireapp/lru-cache@3.0.9) (2018-10-23)
+
+**Note:** Version bump only for package @wireapp/lru-cache
+
+
+
+
+
+## [3.0.8](https://github.com/wireapp/wire-web-packages/tree/master/packages/lru-cache/compare/@wireapp/lru-cache@3.0.7...@wireapp/lru-cache@3.0.8) (2018-10-19)
+
+**Note:** Version bump only for package @wireapp/lru-cache
+
+
+
+
+
+## [3.0.7](https://github.com/wireapp/wire-web-packages/tree/master/packages/lru-cache/compare/@wireapp/lru-cache@3.0.6...@wireapp/lru-cache@3.0.7) (2018-10-12)
+
+**Note:** Version bump only for package @wireapp/lru-cache
+
+
+
+
+
+## 3.0.6 (2018-10-11)
+
+**Note:** Version bump only for package @wireapp/lru-cache
+
+
+
+
+
 <a name="3.0.5"></a>
 ## [3.0.5](https://github.com/wireapp/wire-web-packages/tree/master/packages/lru-cache/compare/@wireapp/lru-cache@3.0.4...@wireapp/lru-cache@3.0.5) (2018-10-01)